+++ R/dfSummary.R
@@ -280,7 +280,7 @@ dfSummary <- function(x,
     }
   }
   
-  if (isTRUE(.st_env$noX11)) {
+  if (isFALSE(st_options("use.x11"))) {
     store_imgs <- FALSE
   } else if (!isTRUE(plain.ascii) && style == "grid" && isTRUE(graph.col)) {
     if (is.na(tmp.img.dir)) {
@@ -525,7 +525,7 @@ crunch_factor <- function(column_data, email_val) {
     counts_props <- align_numbers_dfs(counts, round(props, round.digits + 2))
     outlist[[2]] <- paste0("\\", counts_props, collapse = "\\\n")
     if (isTRUE(parent.frame()$graph.col) && any(!is.na(column_data))) {
-      if (!isTRUE(.st_env$noX11)) {
+      if (isTRUE(st_options("use.x11"))) {
         outlist[[3]] <- encode_graph(counts, "barplot", graph.magnif)
       }
       if (isTRUE(parent.frame()$store_imgs)) {
@@ -570,7 +570,7 @@ crunch_factor <- function(column_data, email_val) {
       tmp_data[which(as.numeric(tmp_data) > max.distinct.values)] <-
         paste("[", n_extra_levels, trs("others"), "]")
       levels(tmp_data)[(max.distinct.values + 2):n_levels] <- NA
-      if (!isTRUE(.st_env$noX11)) {
+      if (isTRUE(st_options("use.x11"))) {
         outlist[[3]] <- encode_graph(table(tmp_data), "barplot", graph.magnif)
       }
       if (isTRUE(parent.frame()$store_imgs)) {
@@ -636,7 +636,7 @@ crunch_character <- function(column_data, email_val) {
     outlist[[2]] <- paste0("\\", counts_props, collapse = "\\\n")
     
     if (isTRUE(parent.frame()$graph.col) && any(!is.na(column_data))) {
-      if (!isTRUE(.st_env$noX11)) {
+      if (isTRUE(st_options("use.x11"))) {
         outlist[[3]] <- encode_graph(c(email_val, dups), "barplot", graph.magnif, 
                                      emails = TRUE)
       }
@@ -664,7 +664,7 @@ crunch_character <- function(column_data, email_val) {
       outlist[[2]] <- paste0("\\", counts_props, collapse = "\\\n")
       if (isTRUE(parent.frame()$graph.col) &&
           any(!is.na(column_data))) {
-        if (!isTRUE(.st_env$noX11)) {
+        if (isTRUE(st_options("use.x11"))) {
           outlist[[3]] <- encode_graph(counts, "barplot", graph.magnif)
         }
         if (isTRUE(parent.frame()$store_imgs)) {
@@ -704,7 +704,7 @@ crunch_character <- function(column_data, email_val) {
         names(counts)[max.distinct.values + 1] <-
           paste("[", n_extra_values, trs("others"),"]")
         counts <- counts[1:(max.distinct.values + 1)]
-        if (!isTRUE(.st_env$noX11)) {
+        if (isTRUE(st_options("use.x11"))) {
           outlist[[3]] <- encode_graph(counts, "barplot", graph.magnif)
         }        
         if (isTRUE(parent.frame()$store_imgs)) {
@@ -748,7 +748,7 @@ crunch_logical <- function(column_data) {
     outlist[[2]] <- paste0("\\", counts_props, collapse = "\\\n")
     if (isTRUE(parent.frame()$graph.col) &&
         any(!is.na(column_data))) {
-      if (!isTRUE(.st_env$noX11)) {
+      if (isTRUE(st_options("use.x11"))) {
         outlist[[3]] <- encode_graph(counts, "barplot", graph.magnif)
       }
       if (isTRUE(parent.frame()$store_imgs)) {
@@ -891,7 +891,7 @@ crunch_numeric <- function(column_data, is_barcode) {
     
     if (isTRUE(parent.frame()$graph.col)) {
       if (length(counts) <= max.distinct.values) {
-        if (!isTRUE(.st_env$noX11)) {
+        if (isTRUE(st_options("use.x11"))) {
           outlist[[3]] <- encode_graph(counts, "barplot", graph.magnif)
         }
         if (isTRUE(parent.frame()$store_imgs)) {
@@ -902,13 +902,13 @@ crunch_numeric <- function(column_data, is_barcode) {
         }
         
         if (isTRUE(extra_space)) {
-          if (!isTRUE(.st_env$noX11)) {
+          if (isTRUE(st_options("use.x11"))) {
             outlist[[3]] <- paste0(outlist[[3]], "\n\n")
           }
           outlist[[4]] <- paste0(outlist[[4]], " \\ \n \\")
         }
       } else {
-        if (!isTRUE(.st_env$noX11)) {
+        if (isTRUE(st_options("use.x11"))) {
           outlist[[3]] <- encode_graph(column_data, "histogram", graph.magnif)
         }
         if (isTRUE(parent.frame()$store_imgs)) {
@@ -954,7 +954,7 @@ crunch_time_date <- function(column_data) {
       props <- round(prop.table(counts), round.digits + 2)
       counts_props <- align_numbers_dfs(counts, props)
       outlist[[2]] <- paste(counts_props, collapse = "\\\n")
-      if (!isTRUE(.st_env$noX11)) {
+      if (isTRUE(st_options("use.x11"))) {
         outlist[[3]] <- encode_graph(counts, "barplot", graph.magnif)
       }
       if (isTRUE(parent.frame()$store_imgs)) {
@@ -991,7 +991,7 @@ crunch_time_date <- function(column_data) {
       
       if (isTRUE(parent.frame()$graph.col)) {
         tmp <- as.numeric(column_data)[!is.na(column_data)]
-        if (!isTRUE(.st_env$noX11)) {
+        if (isTRUE(st_options("use.x11"))) {
           outlist[[3]] <- encode_graph(tmp - mean(tmp), "histogram", graph.magnif)
         }
         if (isTRUE(parent.frame()$store_imgs)) {