+++ R/use_custom_lang.R
@@ -20,15 +20,19 @@
 use_custom_lang <- function(file) {
   
   if (!"file" %in% names(match.call())) {
-    if (interactive() && !isTRUE(.st_env$noX11)) {
+    if (interactive() && isTRUE(capabilities("tcltk"))) {
       file <- character()
-      file <- tclvalue(tkgetOpenFile(initialdir = "~",
-                                     filetypes = "{{csv files} {*.csv}}"))
+      file <- try(tclvalue(tkgetOpenFile(initialdir = "~",
+                                         filetypes = "{{csv files} {*.csv}}")),
+                  silent = TRUE)
+      if (class(file) == "try-error") {
+        stop("Window dialog not permitted; 'file' argument must be specified")
+      }
       if (file == "") {
         stop("operation cancelled")
       }
     } else {
-      stop("'file' argument must be specified")
+      stop("Window dialog not permitted; 'file' argument must be specified")
     }
   }