+++ R/zzz.R
@@ -6,12 +6,7 @@
 # Determine OS : Windows | Linux | Darwin
 .st_env$sysname <- Sys.info()[["sysname"]]
 
-# Check if system has X11 capability on Linux
-# Deprecated for now as it causes problems on some Linux systems with X11 capab.
-# .st_env$noX11 <- Sys.info()[["sysname"]] != "Windows" && 
-#  !isTRUE(capabilities("X11"))
-.st_env$noX11 <- FALSE
-
+# Initialize vector for tempfiles -- useful for cleartmp()
 .st_env$tmpfiles <- c()
 
 # Initialize list used by view() when printing an object of class "by"
@@ -55,7 +50,8 @@ utils::globalVariables(c("."))
                  "dfSummary.silent"       = FALSE,
                  "tmp.img.dir"            = NA_character_,
                  "subtitle.emphasis"      = TRUE,
-                 "lang"                   = "en"))
+                 "lang"                   = "en",
+                 "use.x11"                = TRUE))
 
   return(invisible())
 }
@@ -71,7 +67,8 @@ utils::globalVariables(c("."))
                           "most recent version. You can do so by using \n",
                           "devtools::install_github('rapporter/pander')")
   if (Sys.info()[["sysname"]] != "Windows" && !isTRUE(capabilities("X11"))) {
-    packageStartupMessage("system has no X11 capabilities, therefore only ",
-                          "ascii graphs will be produced by dfSummary()")
+    packageStartupMessage("system might not have X11 capabilities; in case of ",
+                          "errors when using dfSummary(), set ",
+                          "st_options(use.x11 = FALSE)")
   }
 }